+++ src/events/ready.ts
@@ -1,10 +1,9 @@
 import { commandsList } from "@commands";
-import { log, success } from "@console";
+import { log, success, warn } from "@console";
 import type { Settings } from "@interface";
 import { type EClient, VERSION } from "@main";
 import { ActivityType, type Guild, REST, Routes } from "discord.js";
 import dotenv from "dotenv";
-import * as fs from "node:fs";
 import process from "node:process";
 
 dotenv.config({ path: ".env" });
@@ -48,10 +47,10 @@ function convertDatabaseUser(db: Settings, guild: Guild) {
 	const defaultChannel = db.get(guild.id, "managerId");
 	const privateChannel = db.get(guild.id, "privateChannel");
 	for (const [userId, userData] of Object.entries(users)) {
-		console.warn(`Converting ${userId} => ${JSON.stringify(userData)} in ${guild.name}`);
 		for (const index in userData) {
 			const data = userData[index];
 			if (!Array.isArray(data.messageId)) {
+				warn(`Converting ${userId} => ${JSON.stringify(userData)} in ${guild.name}`);
 				let toUpdate = false;
 				if (data.isPrivate && privateChannel) {
 					data.messageId = [data.messageId, privateChannel];
@@ -69,6 +68,7 @@ function convertDatabaseUser(db: Settings, guild: Guild) {
 				}
 			}
 		}
+		if (userData.length === 0) db.delete(guild.id, `user.${userId}`);
 	}
 	db.set(guild.id, true, "converted");
 }
@@ -76,14 +76,10 @@ function convertDatabaseUser(db: Settings, guild: Guild) {
 function cleanData(client: EClient) {
 	const guilds = client.guilds.cache;
 	const settings = client.settings;
-	//remove the guild that the bot is not in anymore
-	const toDelete = [];
 	for (const [guildId] of settings.entries()) {
 		if (!guilds.has(guildId)) {
-			console.log(`${guildId} is not in the guilds list anymore`);
-			toDelete.push(guildId);
+			warn(`Removing ${guildId} from the database as the bot is not in it anymore`);
+			client.settings.delete(guildId);
 		}
 	}
-	const toDeleteStr = `-${toDelete.join("\n- ")}`;
-	fs.writeFileSync("toDelete.txt", toDeleteStr, "utf8");
 }